+++ neural_network/helpers/forward_cache.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from neural_network.math_functions import Array
+from neural_network.network_functions.math_functions import Array
 
 
 @dataclass
