+++ requirements.txt
@@ -1,52 +1,12 @@
-astroid==2.12.13
-attrs==22.2.0
-autoflake==2.0.0
-certifi==2022.12.7
-charset-normalizer==3.0.1
-colorama==0.4.6
 contourpy==1.0.6
-coverage==7.0.3
 cycler==0.11.0
-dill==0.3.6
-exceptiongroup==1.1.0
 fonttools==4.38.0
-future==0.18.2
 h5py==3.7.0
-idna==3.4
-iniconfig==1.1.1
-isort==5.11.4
-joblib==1.2.0
 kiwisolver==1.4.4
-lazy-object-proxy==1.9.0
-mando==0.6.4
 matplotlib==3.6.2
-mccabe==0.7.0
-mypy==0.991
-mypy-extensions==0.4.3
 numpy==1.24.1
 packaging==22.0
-pandas==1.5.2
 Pillow==9.4.0
-platformdirs==2.6.2
-pluggy==1.0.0
-pycodestyle==2.10.0
-pyflakes==3.0.1
-pylint==2.15.9
 pyparsing==3.0.9
-pytest==7.2.0
 python-dateutil==2.8.2
-pytz==2022.7
-PyYAML==6.0
-radon==5.1.0
-requests==2.28.1
-scikit-learn==1.2.0
-scipy==1.10.0
-SCons==4.4.0
 six==1.16.0
-threadpoolctl==3.1.0
-tomli==2.0.1
-tomlkit==0.11.6
-typing_extensions==4.4.0
-urllib3==1.26.13
-wrapt==1.14.1
-xenon==0.9.0
